requirements.txt
Django==5.1
django-extensions==3.2.3
django-hosts==6.0
django-storages==1.14.4
psycopg2-binary==2.9.9
pillow==10.4.0
gunicorn==20.1.0
paramiko==3.4.0
